backend/pyproject.toml
[project]
name = "bfam-rental-api"
version = "0.1.0"
description = "BFam Rental API — FastAPI + Supabase"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2024.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "pyjwt[crypto]>=2.8.0",
    "uvicorn[standard]>=0.32.0",
    "storage3>=0.12.0",
    "supabase>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
